refactor: original, discount, final price representation
Final refactor of price types
- original is int
- discount is BigDecimal
- final is int (rounded)

+++ src/main/java/antigravity/repository/ProductRepository.java
@@ -31,7 +31,7 @@ public class ProductRepository {
         return Product.builder()
                 .id(rs.getInt("id"))
                 .name(rs.getString("name"))
-                .price(rs.getBigDecimal("price"))
+                .price(rs.getInt("price"))
                 .build();
     }
 }
